+++ tests/Steps/Html/XPathQueryTest.php
@@ -2,9 +2,10 @@
 
 namespace tests\Steps\Html;
 
-use Crwlr\Crawler\Steps\Html\CssSelector;
 use Crwlr\Crawler\Steps\Html\XPathQuery;
 use Symfony\Component\DomCrawler\Crawler;
+
+use function tests\helper_getSimpleListHtml;
 
 test('The apply method returns a string for a single match', function () {
     $xml = '<item>test</item>';
@@ -103,4 +104,52 @@ it('turns the value into an absolute url when toAbsoluteUrl() is called', functi
     $query->toAbsoluteUrl();
 
     expect($query->apply($domCrawler))->toBe('https://www.example.com/foo/bar');
+});
+
+it('gets only the first matching element when the first() method is called', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->first();
+
+    expect($selector->apply($domCrawler))->toBe('one');
+});
+
+it('gets only the last matching element when the last() method is called', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->last();
+
+    expect($selector->apply($domCrawler))->toBe('four');
+});
+
+it('gets only the nth matching element when the nth() method is called', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->nth(3);
+
+    expect($selector->apply($domCrawler))->toBe('three');
+});
+
+it('returns null when no nth matching element exists', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->nth(5);
+
+    expect($selector->apply($domCrawler))->toBeNull();
+});
+
+it('gets only even matching elements when the even() method is called', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->even();
+
+    expect($selector->apply($domCrawler))->toBe(['two', 'four']);
+});
+
+it('gets only odd matching elements when the odd() method is called', function () {
+    $domCrawler = new Crawler(helper_getSimpleListHtml());
+
+    $selector = (new XPathQuery("//*[@id = 'list']/*[contains(@class, 'item')]"))->odd();
+
+    expect($selector->apply($domCrawler))->toBe(['one', 'three']);
 });